+++ src/avista_digital_exchange_sdk/__init__.py
@@ -1,4 +1,5 @@
 from .avista_digital_exchange import *
+from . import DxTypes
 # from .data_types import *
 # from .graphql_queries import *
 # from .client import Client as client