+++ src/avista_digital_exchange_sdk/avista_digital_exchange.py
@@ -1,16 +1,14 @@
-import os
+import logging
 from .exceptions import *
 from .iotUtil import *
 from .dataStoreUtil import *
+from .dataCaptureUtil import *
 from .collaborativeUtil import *
 
 from .data_types.user import User
 from .graphql_queries.user_getUserSession import user_getUserSession
 from .client import Client
-from . import globals
 
-
-# from .graphql_queries.storage_getDataStoreFileDownloadUrl import storage_getDataStoreFileDownloadUrl
 
 class AvistaDigitalExchange(object):
 
@@ -18,13 +16,18 @@ class AvistaDigitalExchange(object):
         if type(debug) is not bool:
             raise InvalidParameterException(
                 "AvistaDigitalExchange debug parameter must be a bool")
-        self._stage = "PRODUCTION"
-        self._debug = debug
-        self._token = token
-        self._client = Client(self._token, self._stage, self._debug)
-        self.iot = IoTUtil(self._debug, self._client)
+        self._stage: str = "PRODUCTION"
+        self._debug: bool = debug
+        self._token: str = token
+        self._client: Client = Client(self._token, self._stage, self._debug)
+        self._logger: Logger = logging.getLogger(name="dx-logger")
+        self._logger.setLevel(20)
+        self.iot: IoTUtil = IoTUtil(self._debug, self._client)
         # self.collaboratives = CollaborativeUtil(self._debug, self._client)
-        self.dataStores = DataStoreUtil(self._debug, self._client)
+        self.dataStores: DataStoreUtil = DataStoreUtil(
+            self._debug, self._client)
+        self.dataCapture: DataCaptureUtil = DataCaptureUtil(
+            self._debug, self._client, self._logger)
         print("SDK initialized!")
 
     def getUserInfo(self) -> User:
@@ -34,15 +37,10 @@ class AvistaDigitalExchange(object):
         return result
 
 
-def isValidISO8601Timestamp(timestamp):
+def isValidISO8601Timestamp(timestamp) -> bool:
     import re
 
     matched = re.match("\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z", timestamp)
     isMatch = bool(matched)
 
     return isMatch
-    # try:
-    #     datetime.fromisoformat(timestamp.replace('Z', '+00:00'))
-    # except:
-    #     return False
-    # return True