+++ src/avista_digital_exchange_sdk/graphql_mutations/dataCapture_publishData.py
@@ -1,0 +1,11 @@
+from ..gql2.schema_types import DataCapture_publishDataParams, DataCapture_publishDataMutationResult, Mutation
+
+mut = Mutation["dataCapture_publishData"]
+
+query = QueryType()
+
+
+@query.field('me')
+def resolve_me(obj, info: GraphQLResolveInfo) -> MeQueryResult:
+    #  implementation to obtain current user
+    return mocked_user