+++ tests/test_dataCapture_publishData.py
@@ -1,0 +1,6 @@
+
+from src.avista_digital_exchange_sdk import AvistaDigitalExchange
+
+
+async def runTest(dxInstance: AvistaDigitalExchange):
+    raise Exception("Not implemented")