+++ CHANGELOG.md
@@ -25,6 +25,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-06-15
+
+### Added
+
+- Added new method for generating a graphql client and python types from a graphql schema.
+- Added new Data Capture module with publishData, startCapture, and stopCapture commands.
+
 ## [0.2.1] - 2023-02-15
 
 ### Fixed
@@ -186,8 +193,9 @@ First Beta-ready release.
      - addServiceToCollaborative
      - removeServiceFromCollaborative
 
-[Unreleased]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2023_02_15_v0.2.1...HEAD
-[0.2.0]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2023_02_14_v0.2.0...release/2023_02_15_v0.2.1
+[Unreleased]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2023_06_15_v0.3.0...HEAD
+[0.3.0]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2023_02_15_v0.2.1...release/2023_06_15_v0.3.0
+[0.2.1]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2023_02_14_v0.2.0...release/2023_02_15_v0.2.1
 [0.2.0]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2022_12_25_v0.1.5...release/2023_02_14_v0.2.0
 [0.1.5]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2022_12_21_v0.1.4...release/2023_01_25_v0.1.5
 [0.1.4]: https://github.com/Avista-Digital-Innovation/avista-digital-exchange-sdk/compare/release/2022_12_19_v0.1.3...release/2022_12_21_v0.1.4